[jest] revert spyOn overload breaking change
PR #6278 introduced a breaking change to the jest types for spyOn by
changing the overload ordering. This undoes that change.

An example that was affected by the breaking change is added to the
tests.

Reverting the ordering broke one of the tests that was added in #6278.
However, that test was ambiguous because `Date` is function in addition
to a class. Replacing it with a class that requires the `new` keyword
like `Error` passes the test.