Use ModelSchema in ember-data adapter/serializers
Ember Data passes a private
[ShimModelClass](https://github.com/emberjs/data/blob/3aa0fbddb7a650361bed44350c3b51669ec9065e/packages/store/addon/-private/legacy-model-support/shim-model-class.ts#L38)
as the `type/modelClass/primaryModelClass` argument to many of the
public methods, as opposed to an instance of the model, or even the
class itself. The fallout is that the current types don't match the implementation,
inclugin flagging any usage of `modelName`, a static property on the
class.

This adds a `ModelSchema` interface, which matches the private, in-repo
type implemented by `ShimModelClass`. It also updates the call sites
with the new type, where appropiate.