Ember: backport Owner improvemenets from Ember's own types
This should let apps and libraries which are not yet ready to switch
over to Ember's preview/stable types nonetheless be able to rely on the
improvements we landed in these APIs, and aligns them to minimize the
breaking changes from the DT types to Ember's table types on these
particular fronts. Basically, this unlocks letting libraries use DT
*or* Ember's own types for these during the transition, with them
provided as appropriate by the host environment. Painful, but useful.