Fix fs-extra usage in scripts
`fs-extra` was not added as a dependency. Scripts that were using `fs-extra` relied on an indirect dependency on an
older version of `fs-extra`. These scripts only used functions that `fs-extra` exports from `fs`. The latest version of
`fs-extra` doesn’t even export these in the ESM export, only in the CJS export. So while my original idea was to update
to the latest version of `fs-extra`, instead I removed the dependency entirely.