mongo: Fix types of iterators and dropIndexAsync
Declaring iterators and async iterators with a "next" type of never
makes them unusable with TypeScript, which throws an error:

> Cannot iterate value because the 'next' method of its iterator expects
> type 'never', but for-of will always send 'undefined'.ts(2763)

However, Meteor's iterators are agnostic to the value passed in, so
leaving it with the default of undefined is reasonable.

Also, dropIndexAsync was missing a return type (it should return a
promise, not void).