Added JSDoc from upstream for intelliSence.
Reason for changes:

	The in-editor IntelliSence for the Passport.js (`passport`) library was not
displaying the JSDoc comments present in the passport source code.  So
while the @types/passport package added the typescript typings for passport, it
removed the ability to see the documentation JSDoc comments.

Changes:

In types/passport/index.d.ts:

	Added the JSDoc comments from the upstream
https://github.com/jaredhanson/passport repo to get the
documentation only found in the source code JSDoc comments.  With
the JSDoc comments for the .d.ts types, the integrated JSDoc
IntelliSence viewer for most editors will now display the JSDoc
documentation correctly.  Also added a missing typing for the built-in
passport authentication strategy `SessionStrategy`, which was not
present previously.  Also updated some missing options for some
functions according to the documentation.

Need for review:
	The `SessionStrategy.authenicate` method is a protected member
of the `SessionStrategy` class.  However, I could not find a way to
properly express that type while using the typescript `interface`, as
it seems the only typescript classes that have the ability to use the
`protected` modifier.  So I used a JSDoc annotation instead. Is using
JSDoc annotations acceptible in this repo?

Request for maintainers:
	Please see if it is reasonable to keep the JSDoc comments for
the Passport.js (`passport`) library package up to date with the upstream
documentation.  I would be open to see how I could assist in this regard.