Clone only the needed parts of aws-sdk@2
aws-sdk@2 is one huge package that has since been broken into multiple
packages in the @aws-sdk scope starting with version 3. Unfortunately,
dtslint-runner often runs out of memory compiling dependents of
aws-sdk@2 because it's so large. This despite the fact that
dependents of aws-sdk@2 on DT about 5 parts of the SDK.

This PR copies aws-sdk@2.1303.0. It includes the top-level of lib/ and a
few subdirectories. From clients/ it includes

- dynamodb
- rdsdataservice
- s3
- ses
- sqs
- ssm

Then it updates the following packages to use aws-sdk2-types instead of
aws-sdk:

- amazon-dax-client
- data-api-client
- lesgo
- nodemailer-ses-transport
- s3-streams
- aws-param-store
- dynogels
- s3-download-stream
- s3-upload-stream

These packages go from requiring 800+ MB to compile to 200 MB.
Compilation speed is much faster too.

If the types need to be updated the required files can be copied
from aws-sdk again.