[express-session]: expand some existing types to match implementation
When a cookie is constructed without arguments, the `expires` and
`originalMaxAge` fields are null. Existing tests within the
`express-session` repo reflect this, along with other manual tests.

Additionally, the `Store` class's `length` method will conditionally
call the given callback without the second `length` parameter when there
is an error. This parameter has been made optional to reflect that.

As indicated by test failures in an earlier version of this change, an
equivalent change was required in the `connect-sqlite3` types package.
It appears as though there are some duplicate types that could be
consolidated. If so, that would belong in a different commit.