Nightwatch: delete unused jsdoc types
Previously, the lint rule forbidding types in jsdoc was crashing.
Now that it's not, it found some types in nightwatch's jsdoc.
After the types are gone, there's not really any additional text in the param tags, so I deleted them, too.
The whole thing is redundant with the type annotations.