[README]: Improve Chinese translation (#64116)
* [README]: rename zh.md to zh-Hans.md

Using zh is ambiguous because it does not tell it is Simplified
Chinese or Traditional Chinese. As this translation is actually
in Simplified Chinese, so rename it.

* [README]: improve Simplified Chinese translation

* [README]: adopt translation from current zh docs

PR #63899 updated the Chinese translation and was merged. I took
a look at their work and adopted some of their tranlations that
are better than mine.

* [README]: Update links in es.md and pt.md

Also change title of link to zh-Hans.md from "Chinese" to "Simplified
Chinese".

* [README]: Sync with latest English readme