Update dom-screen-wake-lock for TS 5.1
Typescript 5.1's DOM types add the same types as dom-screen-wake-lock.
This PR updates the types to make them agree so that people can keep
installing (the latest version of) this package.

However, I don't think this package will actually be needed after 5.1,
because all of its types are now in the standard library.