Add missing manifest permission types
"downloads.shelf" has been around for a while. It is not documented on
the main list of permissions
<https://developer.chrome.com/docs/extensions/reference/downloads/>, but
is described on the reference page for the "downloads" API
<https://developer.chrome.com/docs/extensions/reference/downloads/>.

"downloads.ui" is new in Chrome 105. Like "downloads.shelf", it is only
documented on the reference page for the "downloads" API.

"offscreen" is new in Chrome 109.