Allow jsonwebtoken to use empty key with algorith none.
Algorithm none is (of course) insecure, but it is a part of JWT standard.
I am using it in tests, to test that my server doesn't let an unsigned JWT in.

I was thinking about also adding it somehow to verify, but that seemed a little harder and honestly shouldn't be really used :)