feat(itemsjs): Fix incorrect key type on buckets
Closes #64945

The key type for buckets should be its possible values based on an
aggregation, not the name of the key itself. For example, if an
aggregation is declared for the key `foo`, then the `key` field on
`Bucket` should be the possible values of `foo`.

This PR also improves the type requirement on the aggregation generic
(`A`) to ensure that it's a key of the item type `I`.