Remove global JSX namespace
The goal of just adding the scoped namespace
was to ensure the ecosystem  can prepare.
However, you'd still have to change augmentations to target the scoped namespace
i.e. making this a breaking change.
So if we break anyway, we might break everything. The additional breakage is
codemoddable after all.