fix conflicting intersection for ExecFileException
The error parameter of `execFile`'s callback has a property `code`, which can have various types:
- string when a system call failed (for example: ENOENT)
- number when the child process exited with a non-zero status code
- null when killed

However, `ExecFileException` in the previous type definition has `code?: undefined` because it is an intersection of `code?: number | undefined` from `ExecException` and `code?: string | undefined` from `NodeJs.ErrnoException`.

To fix this, `ExecFileException` is now an intersection of `ExecException` and `NodeJs.ErrnoException` without the `code` property, and a dedicated object type for the `code` property.

I haven't found conditions under which the `code` property is missing or the value is undefined. However, I defined its type as `code?: string | number | undefined | null` in order to make it compatible with the previous definition `code?: undefined`.