nunjucks: fix async ILoader and increase type-safety
While the ILoader interface previously attempted to encompass both
synchronous and asynchronous loaders, it actually failed to encompass
the latter. Attempting to implement an async ILoader:

    class MyLoaderAsync extends nunjucks.Loader implements nunjucks.ILoader {
        async: true;
        getSource(name: string, callback) {}
    }

previously failed with the following type error:

    Property 'getSource' in type 'MyLoaderAsync' is not assignable to the same property in base type 'ILoader'.
      Type '(name: string, callback: any) => void' is not assignable to type '{ (name: string): LoaderSource; (name: string, callback: Callback<Error, LoaderSource>): void; }'.
        Target signature provides too few arguments. Expected 2 or more, but got 1.

(Because the ILoader interface previously declared both the signatures
for a synchronous getSource and for an asynchronous getSource via
overloading, which doesn't make sense ... it's either or, not both.)

The previous type definitions also failed to detect mismatches
between the async boolean and the getSource implementation.

This commit fixes the first bug by reducing the ILoader to synchronous
loaders (and introducing an ILoaderAsync interface for asynchronous
loaders) and the second deficiency by introducing a stricter union.