Fix katex exports
Katex uses UMD for its exports. It also has an ESM entry point which exposes the same thing using a default export, so
that’s compatible with the CJS entry point as-is. `katex/contrib/auto-render` is also a UMD module.

There are more exports in the `katex/contrib/*` folders, but `katex2tex` is not one of them, so its types were deleted.

The ESLint violations were resolved, so `.eslintrc.json` could be deleted.