[node] corrects test event duration_ms name
When the Node test runner passes or fails a test, it sends an event to
registered reporters. The event has details about the test run,
including the duration. The property name of the duration has always
been duration_ms, but the documentation had it listed as just duration.
That has recently been fixed (https://github.com/nodejs/node/commit/71165e111a8606ad37784d992b8d3cee4fbf0156).

This updates the types to be correct.

This also adds a test reporter so that tests can be written for the
reporter events' types.