[node] adds "type" to test runner event details
In [version 20.0.0 (and backported to 19.9.0 and 18.7.0)](https://github.com/nodejs/node/pull/47094)
the test runner started reporting on whether a test was a suite. This
was exposed to reporters in the `details` object of a `test:pass` or
`test:fail` event but this hasn't been documented. This adds the `type`
property to both event's `details` object.