Refactor class defs to interfaces and unexport
We should work with interface types instead of classes, since working
with a class infers that the type can be instantiated with the new
operator. With the k6 browser module we currently do not need to new
up an instance of any type that the module works with.

We need to correctly reflect that the k6 browser module only exports
browser and not the other types that were being exported. This also
does a better a job of showing us how to get to some of the other types
when working with browser e.g. page, response, request etc.