remove(@hapi/nes): remove just sources of the package for #66292
This is technical removal. In #66292 there is an effort to update DT for
upcoming TS language changes (see linked PR for details) and updating
this specifc DT package (@hapi/nes) to pass checks againsts current
status of TS language would prove fruitless. The upstream JS package
already progressed and DT types are out of sync with current module.
One could just revert this PR, in case of change to @hapi/nes types is
to be done in future.

Thank you for your understanding!