Add new aws-sns package
As part of !66410 it was discovered that the json payload send to aws lambda
does not match the JSON send by SNS to other custom HTTP/HTTPS endpoints or
services. This new package aims to better serve this use case of wanting to
type AWS/Amazon Simple Notification Service (Amazon SNS) payloads that are NOT
used with aws lamda. Having a dedicated package will also discoverability.