[node]: Fix test runner HookFn method signature
This fixes an incorrect method signature on the `HookFn` type used by
hooks such as `before` and `after`. Per the Node documentation, and
based on my experience and observations using the Test Runner, the
callback function is the second argument to the hook function. See
https://nodejs.org/api/test.html#beforefn-options.

The first argument is not documented but it appears to be a
`SuiteContext` object like the first argument passed to `describe`.