fix(dcp-client): add exported wallet module symbols
The Wallet API from `dcp-client` also exports constructors for `Address`
and `PrivateKey`, so types have been added to indicate that they're
exposed.

The need to `await` the `Keystore` constructor is now also expressed
properly in the types.

The tests for the global types has been minimized since the key part of
the test is `window.dcp` & it's properties' types. As long as those are
correct, the module tests should cover the deeper types.

Refs: https://stackoverflow.com/a/71513155