[dns2] Add support for IPv6 UDP DNS servers
To instantiate a UDP IPv6 DNS server, it is necessary to pass
an object with type set as 'udp6' as a constructor argument.
This will cause dns2 to open the right kind of socket.

See https://github.com/song940/node-dns/blob/master/server/udp.js

This commit updates the UDP objects to add this facility.

The address has been made optional in the UDP listen signature
to allow the socket to choose the right all interface address.