fix(bn.js): add missing second argument to setn
bn.js in version 5.1 accepts two arguments for setn function
(bit index and value to use), but README.md listed wrong signature
for this function so this issue carried to this package.
https://github.com/indutny/bn.js/commit/6ef485dc30b002ec8100602998763fec8a1a3897

bn.js only checks if value is truthy or not to decide what to do, so it could really
be anything. I set it to boolean | 0 | 1 as those make sense for this function.