[react-instantsearch-core] Explicitly rely on `ReactElement<any>` in tests
In React 19, we'll default `element.props` to `unknown`. However, your tests rely on the old behavior of defaulting to `any`.

To reduce the size of the React 19 types PR, I'll encode this behavior now.

Feel free to adjust your tests to leverage `unknown` now in a follow-up by explicitly using `ReactElement<unknown>`.