Allow mocha callbacks to use any subclass of Context
When creating fixtures in Mocha, they get attached to Context, which includes an index signature for this purpose.  However, those fixtures will have no type information.

If a project wants to strongly type its fixtures, it needs to subclass Context.  But then, without this patch, the compiler will reject the use of those subclasses in Mocha callbacks.