feat(jwplayer): default and type are not required
**default**

Set this to `true` for the media source you want to play on startup. If this isn't set for any `source`, the first one is used

**type**

Forces a media type. Only required when a file extension is missing or not recognized (Using .php or certain tokens, for example)

See https://docs.jwplayer.com/players/reference/playlists#quality-settings-for-video-files