Make ImportMeta.resolve conditional
The DOM type for ImportMeta.resolve is `(specifier: string): string`,
and Typescript 5.6 will ship this type, which conflicts with the type
in node/v16 and node/v18.

This PR makes the type conditional -- when the DOM is present, it will
use the DOM type. When it's not, it will use the type as it is today.

I had to cut the tests down so they would pass in either case.