[react] Explicitly allow async Form Actions
This was implicitly enabled by TypeScript allowing any return type if the return type of an expected function is `void`.
But this is not generally behavior we want in React so we should not relly on it.

It's also problematic since certain type-aware ESLint rules may issue violations since it looks like Form Actions can't be async.

We now explicitly codify support of async Form Actions. This also means return anything else will now issue type errors