feat(@types/react): trusted types for ScriptHTMLAttributes, ReactNode, and other elements
Changelog:

- [ ] ScriptHTMLAttributes: src attribute now accepts TrustedScriptURL.
- [ ] ReactNode now accepts TrustedScript
- [ ] SVGAttributes: href and xlinkHref attributes now accept
  TrustedScriptURL
- [ ] IframeHTMLAttributes: srcDoc attribute now accepts TrustedHTML

Rationale:

In order to support trusted types for some of our Next.js applications,
we've had to patch `@types/react` locally to use these types. We're
hoping these will be useful to the broader community.

References:

The types are based on the table found in the Trusted Types W3C specification:

- [ ] https://www.w3.org/TR/trusted-types/#enforcement-in-scripts
- [ ] https://www.w3.org/TR/trusted-types/#get-trusted-type-data-for-attribute