k6: move to use the module syntax
This is done so we can actually define and use ESM specific features such as `import.meta`, top-level-await and in the
future dynamic imports. (And probably more)

The majority of the changes exists to appease typescript to have the same imports.

Arguably this new folder structure more closely represents the reality.

There unfortunately are likely some problems as moving between script and module changes how typescript behaves in more
than one way.