Swap function signatures in Backbone types
**Why**

TypeScript uses the first type signature that matches. In this case,
Array is a sub-class of `{}`, so it thinks `add` will always return a
single model. Swapping the order makes it so that TypeScript correctly
expects the function to return an array of models when it is given an
array.