oauth2-server: fix AbstractGrantType class
The abstract grant type was missing a model in its constructor options,
but excluding one caused an error in the JavaScript.

Furthermore, BaseModel.saveToken does NOT use the client and user
within the token, instead using separate parameters for these,
so I switched the argument to a PartialToken type omitting them,
to avoid confusion and so the upstream example code passes type checks.

The test is updated with example code from the library's documentation on
Extension Grants.

I also added a Scope type to simplify things.