Allow overriding Backbone model with a function
The `model:` property on Backbone Collections defaults to a constructor.
However, they also allow for it to be a function that [dynamically
returns an instance][dy]. Unfortunately, due to limitations in
TypeScript this [cannot be defined as a method][me]. However, we can
assign it as an arrow function instead.

[dy]: http://backbonejs.org/#Collection-model
[me]: https://github.com/Microsoft/TypeScript/issues/27965