rax: fix JSX RaxFragment usage/export
Typescript 5.7 will fail unless usage of RaxFragment is explicit, with
an import of it and a `/** @jsxFrag RaxFragment */` pragma at the top of
the file. This PR adds that.

RaxFragment also needs to have an exported value. I wasn't sure what to
use here so I copied preact, since preact is used in the Typescript
handbook as an example:

```ts
export const RaxFragment: FunctionComponent<{}>
```

Notably, though, the preact source has a comment indicating that
*they're* not sure of the correct type either.

Here is the Typescript 5.7 PR that made this error appear:

https://github.com/microsoft/TypeScript/pull/59933