Add Cast owners
@thibauts is the author of the castv2 package, and so should have
ownership rights over the TS types for that package.

@juliandomingo, @joeyparrish, and @theodab are members of the
Chromecast team at Google, who are responsible for the protocol
(which you can interact with via the third-party castv2 package) and
the SDKs (represented by types in chromecast-caf-receiver and
chromecast-caf-sender), and so should have ownership rights over the
types of all three packages.

We will try to keep APIs and protocol messages here in sync with
upstream changes in the future.