Remove ambient module declaration of @wordpress/blocks
As suggested by the overridden lint rule, a module declaration for the current
package is an antipattern that prevents consumers from declaring the module
themselves to provide new types:

> [T]he declarations in an augmentation are merged as if they were declared in
> the same file as the original. [...] However ... [y]ou can’t declare new
> top-level declarations in the augmentation — just patches to existing
> declarations.

https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation

This ambient module declaration provides no value itself and the effect is the
same when it is omitted. This also matches the approach taken in other
`@wordpress` DefinitelyTyped packages.