[@types/node] correct `ResolveFnOutput.format` (can be an intermediary value)
`format` can be any string, it does not have to be one of the final values. It was designed that way so it can signal to its partner `load` hook that the current module is relevant. ex resolve's `format` can be `typescript`, and then `load` checks `format === 'typescript` to know whether it do anything.

Only `LoadFnOutput.format` is restricted to `ModuleFormat`.