Update tests in react-plotly.js to be forward-compatible with Plotly v3
Long ago it used to be the standard in Plotly to have separate props for `title` (as a string), `titlefont`, etc. in places like the main plot layout and axis definitions.  This was deprecated a while back in favour of a structured object value for `title` with `text`, `font`, etc. as sub-properties.

The old style has been removed entirely in Plotly v3, so this commit updates the _tests_ in react-plotly-js to be forward compatible, ready for when the types are updated to v3.