fix(node): loosen type of LoadFnOutput
A `LoadHook` accepts a `LoadHookContext` object having a `format` of `string | null | undefined`.

`LookHook` implementors often want to pass this value through ([example from docs](https://nodejs.org/api/module.html\#asynchronous-version), which suggests to me that a) it can be falsy, and b) it should be a looser type as established for `ResolveFnOutput` in #71493).