Add extended types to columnStyles/rowStyles
columnStyles and rowStyles both had CellStyle as a possible type, however there are extended properties for each of those (width/height for columns/rows respectively). This seemed to be started as their was a ColumnOptions interface, however it wasn't actually in use, and RowOptions existed but had a lot more options than likely needed. Cleaned it up, seems to be working (allowing me to use minHeight property in my table rows now, huzzah!)