make webpack a peerDependency of @types/webpack for better deduplication
I was upgrading webpack in my repo yesterday, and discovered the hard way that I also had to explicitly upgrade the transitive webpack dependency of @types/webpack to avoid extra type errors.  if webpack was a peerDependency of @types/webpack, I think we'd get this deduplication for free - as it should ~always be the case that anyone using @types/webpack is probably also using webpack.   (if there was a better way to say "no really please deduplicate this" I'd rather use that, but I think peerDependencies the current best way to message that to the package managers)