[passport-http-bearer] make done callback `info.scope` property optional
This allows end users to pass arbitrary data on this object, while still
providing some hints about the “normal” pattern described (but not
required!) by the docs. The implementation simply passes through the
values on this object to be used by Passport as values on the `authInfo`
object, without even inspecting them.