fix(culori): Allow passing mode: null to mapper
According to the culori documentation: "The mode parameter can be set to null, in which case the mapper will iterate through the channels in the color's original color space.".
Passing `null` to `mode` should now be allowed, and will return a color in the same mode that was passed in.