[jsdom] Add type for VirtualConsole.forwardTo in v27
Version 27 of jsdom was released. Included in this release is a change
to the VirtualConsole API summarized in pull request:
https://github.com/jsdom/jsdom/pull/3866

The new API is documented in the README under the section "Virtual
consoles".

https://github.com/jsdom/jsdom#virtual-consoles

The version in package.json has been bumped to reflect the introduction
of the changed API.

The complete release notes for v27 are available here:
https://github.com/jsdom/jsdom/releases/tag/27.0.0