Update type declarations for gulp-sass v6
The new version defaults to using the modern Sass JS API instead of the
legacy ones, while still providing a legacy API using the Sass JS legacy
API for projects not able to migrate yet.

The new type declarations are also stricter about the API, by defining
the expected type for the compiler during initialization instead of
using `any`.
A trick is used to support the `NodePackageImporter` where each
implementation supports only its own implementation of it, allowing both
to support `sass` and `sass-embedded` at the same time in type
declarations and to enforce using the correct implementation, as the
`NodePackageImporter` from `@sass/types` is treated as incompatible with
the `NodePackageImporter` from `sass` or `sass-embedded` right now.